Rename core file
