+++ artistwordcloud/cloud_creation.py
